Fix working on first install
Turns out that shipped extensions behave differently, and the previous solution
of using onInstalled, doesn't work in the wild. Change the Firefox version to
use the management.onEnabled event instead.

Also, change the Chrome version to be mostly like the Firefox one, running in
spanning instead of split incognito mode.