Fix NID cookie sometimes not being found
In a previous version we registered an onHeadersReceived listener just-in-time
for each request. Turns out that on Chrome (130.0.6723.116) the callback will
sometimes be called without the set-cookie headers, even though they were sent
by the server. Interestingly, adding another listener that always exists fixes
it. I suspect there is a race, where the "extraHeaders" property isn't yet
applied to the listener when the response is received. The current version uses
a single global listener to avoid all those problems.